requirements.txt
matplotlib==3.6.3
numpy==1.23.5
#Pillow==9.3.0
Pillow==10.1.0
scipy==1.9.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
